+++ vanilla/items.js
@@ -1,0 +1,245 @@
+import { Blocks, Items, Item } from 'definitions'
+import { itemify, slabifyItem } from './blockshapes.js'
+import {terrainPng} from './blocks.js'
+
+const {Texture} = loader(import.meta)
+
+export const itemsPng = Texture("items.png")
+
+Items.sand = class extends Item{
+	places(){ return Blocks.sand }
+	static texture = Blocks.sand.texture
+	static defaultName = 'Sand'
+}
+
+Items.oak_log = class extends Item{
+	places(){ return Blocks.oak_log }
+	static texture = Blocks.oak_log.texture
+	static defaultName = 'Oak log'
+}
+Items.birch_log = class extends Item{
+	places(){ return Blocks.birch_log }
+	static texture = Blocks.birch_log.texture
+	static defaultName = 'Birch log'
+}
+Items.spruce_log = class extends Item{
+	places(){ return Blocks.spruce_log }
+	static texture = Blocks.spruce_log.texture
+	static defaultName = 'Spruce log'
+}
+Items.dark_oak_log = class extends Item{
+	places(){ return Blocks.dark_oak_log }
+	static texture = Blocks.dark_oak_log.texture
+	static defaultName = 'Dark oak log'
+}
+Items.acacia_log = class extends Item{
+	places(){ return Blocks.acacia_log }
+	static texture = Blocks.acacia_log.texture
+	static defaultName = 'Acacia log'
+}
+Items.jungle_log = class extends Item{
+	places(){ return Blocks.jungle_log }
+	static texture = Blocks.jungle_log.texture
+	static defaultName = 'Jungle log'
+}
+
+Items.oak_planks = class extends Item{
+	places(){ return Blocks.oak_planks }
+	static texture = Blocks.oak_planks.texture
+	static defaultName = 'Oak planks'
+}
+Items.birch_planks = class extends Item{
+	places(){ return Blocks.birch_planks }
+	static texture = Blocks.birch_planks.texture
+	static defaultName = 'Birch planks'
+}
+Items.spruce_planks = class extends Item{
+	places(){ return Blocks.spruce_planks }
+	static texture = Blocks.spruce_planks.texture
+	static defaultName = 'Spruce planks'
+}
+Items.dark_oak_planks = class extends Item{
+	places(){ return Blocks.dark_oak_planks }
+	static texture = Blocks.dark_oak_planks.texture
+	static defaultName = 'Dark oak planks'
+}
+Items.acacia_planks = class extends Item{
+	places(){ return Blocks.acacia_planks }
+	static texture = Blocks.acacia_planks.texture
+	static defaultName = 'Acacia planks'
+}
+Items.jungle_planks = class extends Item{
+	places(){ return Blocks.jungle_planks }
+	static texture = Blocks.jungle_planks.texture
+	static defaultName = 'Jungle planks'
+}
+
+Items.oak_planks_slab = slabifyItem(Items.oak_planks, Blocks.oak_planks)
+Items.birch_planks_slab = slabifyItem(Items.birch_planks, Blocks.birch_planks)
+Items.spruce_planks_slab = slabifyItem(Items.spruce_planks, Blocks.spruce_planks)
+Items.dark_oak_planks_slab = slabifyItem(Items.dark_oak_planks, Blocks.dark_oak_planks)
+Items.acacia_planks_slab = slabifyItem(Items.acacia_planks, Blocks.acacia_planks)
+Items.jungle_planks_slab = slabifyItem(Items.jungle_planks, Blocks.jungle_planks)
+
+Items.sandstone = class extends Item{
+	places(){ return Blocks.sandstone }
+	static texture = Blocks.sandstone.texture
+	static defaultName = 'Sandstone'
+}
+Items.stone = class extends Item{
+	places(){ return Blocks.stone }
+	static texture = Blocks.stone.texture
+	static defaultName = 'Stone'
+}
+Items.glass = class extends Item{
+	places(){ return Blocks.glass }
+	static texture = terrainPng.at(1, 3)
+	static defaultName = 'Glass'
+}
+Items.bedrock = class extends Item{
+	places(){ return Blocks.bedrock }
+	static texture = Blocks.bedrock.texture
+	static defaultName = 'Bedrock'
+}
+Items.obsidian = class extends Item{
+	places(){ return Blocks.obsidian }
+	static texture = Blocks.obsidian.texture
+	static defaultName = 'Obsidian'
+}
+Items.glowing_obsidian = class extends Items.obsidian{
+	places(){ return Blocks.glowing_obsidian }
+	static texture = Blocks.glowing_obsidian.texture
+	static defaultName = 'Glowing obsidian'
+}
+Items.netherrack = class extends Item{
+	places(){ return Blocks.netherrack }
+	static texture = Blocks.netherrack.texture
+	static defaultName = 'Netherrack'
+}
+Items.grass = class extends Item{
+	places(){ return Blocks.grass }
+	static texture = Blocks.grass.texture
+	static defaultName = 'Grass block'
+}
+Items.dirt = class extends Item{
+	places(){ return Blocks.dirt }
+	static texture = Blocks.dirt.texture
+	static defaultName = 'Dirt'
+}
+Items.sugar_cane = class extends Item{
+	places(){ return Blocks.sugar_cane }
+	static texture = itemsPng.crop(16,256,16,16)
+	static defaultName = 'Sugar cane'
+}
+
+class Tool extends Item{
+	static maxStack = 1
+	static model = 1
+	static for = ''
+	static speed = 10
+	breaktime(block){
+		return block.tool == this.for ? block.breaktime / this.speed : block.breaktime
+	}
+}
+
+Items.diamond_pickaxe = class extends Tool{
+	static for = 'pick'
+	static speed = 25
+	static texture = itemsPng.at(2, 8)
+	static defaultName = 'Diamond pickaxe'
+}
+
+Items.diamond_shovel = class extends Tool{
+	static for = 'shovel'
+	static speed = 25
+	static texture = itemsPng.at(3, 8)
+	static defaultName = 'Diamond shovel'
+}
+
+Items.diamond_axe = class extends Tool{
+	static for = 'axe'
+	static speed = 25
+	static texture = itemsPng.at(14, 7)
+	static defaultName = 'Diamond axe'
+}
+
+Items.flint_and_steel = class extends Tool{
+	static model = 2
+	static texture = itemsPng.at(17, 10)
+	static defaultName = 'Flint and steel'
+}
+
+Items.tnt = class extends Item{
+	static texture = Blocks.tnt.texture
+	places(){ return Blocks.tnt }
+	static defaultName = 'TNT'
+}
+
+Items.end_crystal = class extends Item{
+	static texture = itemsPng.at(0,10)
+	static defaultName = 'End crystal'
+}
+class Mineral extends Item{}
+
+Items.lapis = Mineral
+Items.coal = Mineral
+Items.iron = Mineral
+Items.gold = Mineral
+Items.emerald = Mineral
+Items.diamond = Mineral
+
+Items.cut_sandstone = class extends Items.sandstone{
+	static defaultName = 'Cut sandstone'
+}
+Items.smooth_sandstone = class extends Items.sandstone{
+	static defaultName = 'Smooth sandstone'
+}
+Items.chiseled_sandstone = class extends Items.sandstone{
+	static defaultName = 'Chiseled sandstone'
+}
+Items.red_sandstone = class extends Items.sandstone{
+	static defaultName = 'Red sandstone'
+}
+Items.cut_red_sandstone = class extends Items.sandstone{
+	static defaultName = 'Cut red sandstone'
+}
+Items.chiseled_red_sandstone = class extends Items.sandstone{
+	static defaultName = 'Chiseled red sandstone'
+}
+Items.smooth_red_sandstone = class extends Items.sandstone{
+	static defaultName = 'Smooth red sandstone'
+}
+
+Items.end_portal_frame = class extends Item{
+	static texture = terrainPng.at(9, 0)
+	places(){ return Blocks.end_portal_frame }
+	static defaultName = 'End portal frame'
+}
+
+Items.eye_of_ender = class extends Item{
+	static texture = itemsPng.at(1, 10)
+	static defaultName = 'Eye of ender'
+}
+
+Items.endstone = class extends Item{
+	static texture = terrainPng.at(7, 4)
+	places(){ return Blocks.endstone }
+	static defaultName = 'Endstone'
+}
+
+Items.white_wool = itemify(Blocks.white_wool, 'White wool')
+Items.light_grey_wool = itemify(Blocks.light_grey_wool, 'Light grey wool')
+Items.grey_wool = itemify(Blocks.grey_wool, 'Grey wool')
+Items.black_wool = itemify(Blocks.black_wool, 'Black wool')
+Items.red_wool = itemify(Blocks.red_wool, 'Red wool')
+Items.orange_wool = itemify(Blocks.orange_wool, 'Orange wool')
+Items.yellow_wool = itemify(Blocks.yellow_wool, 'Yellow wool')
+Items.lime_wool = itemify(Blocks.lime_wool, 'Lime wool')
+Items.green_wool = itemify(Blocks.green_wool, 'Green wool')
+Items.cyan_wool = itemify(Blocks.cyan_wool, 'Cyan wool')
+Items.light_blue_wool = itemify(Blocks.light_blue_wool, 'Light blue wool')
+Items.blue_wool = itemify(Blocks.blue_wool, 'Blue wool')
+Items.purple_wool = itemify(Blocks.purple_wool, 'Purple wool')
+Items.magenta_wool = itemify(Blocks.magenta_wool, 'Magenta wool')
+Items.pink_wool = itemify(Blocks.pink_wool, 'Pink wool')
+Items.brown_wool = itemify(Blocks.brown_wool, 'Brown wool')